tests/reproducible_build: fix vmlinux finding logic
In updating the logic that finds the path for vmlinux, it is useful
to support a few different options

1. command-line parameter specification
2. retrieval via pahole option
3. environment variable based retrieval

Options 2 and 3 are useful for the case where the script is run
via the tests script.  When adding support for 3 I incorrectly

vmlinx=${vmlinux:-1}

...which should have been

vmlinux=${vmlinux:-$1}

i.e. assign var vmlinux the the env var value vmlinux, and if that
is not set, use the first parameter for the vmlinux value.  If neither
are set we can then fall back to the pahole option.

Fixes: 834414169572 ("tests/reproducible_builds: Support envvar-specified vmlinux")
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>