btf_encoder: Allow encoding VARs from many sections
Currently we maintain one buffer of DATASEC entries that describe the
offsets for variables in the percpu ELF section.

In order to make it possible to output all global variables, we'll need
to output a DATASEC for each ELF section containing variables, and we'll
need to control whether or not to encode variables on a per-section
basis.

With this change, the ability to emit VARs from multiple sections is
technically present, but not enabled, so pahole still only emits percpu
variables.

A subsequent change will enable emitting all global variables.

Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Link: https://lore.kernel.org/r/20241004172631.629870-4-stephen.s.brennan@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>