tests: Use PAHOLE_LANG_EXCLUDE=rust to skip Rust CUs
Until we have better support for Rust in both the DWARF loader and in
the BTF encoder we better just skip them to avoid needless warnings
_and_ to test the language exclusion logic that is used in the Linux
kernel build.

As per review/test from Alan Maguire, use 'export' to set
PAHOLE_LANG_EXCLUDE, not doing so would only set it as a shell variable,
not adding it to the environment that is used by the subprocesses in the
test/ shell scripts.

I must have thought it was working because while testing I had it
exported, duh.

Cc: Alan Maguire <alan.maguire@oracle.com>
Cc: Andrii Nakryiko <andrii.nakryiko@gmail.com>
Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Miguel Ojeda <ojeda@kernel.org>
Cc: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>