btf_loader: Support BTF encoding from (multiple) BTF object files
pahole separates debug info format reading into loaders for DWARF,
BTF.  Currently we load from DWARF to convert to BTF, but in
the future it will be beneficial to support loading gcc-
(or clang-) generated BTF from objects to be deduplicated and
converted to BTF; in essence taking multiple sources of BTF
or a linker-deduplicated BTF archive and post-processing it.

To get there, a few issues need to be fixed in the BTF loader;
the function prototype representation needs to be modified to
fit with the DWARF-generated one and we need access to ELF
information for objects when loading BTF for BTF encoding.

With these changes in place it becomes possible to load
BTF from multiple objects and generate a detached BTF
output; for example I built a kernel with the gcc "-gbtf"
option, and running for example

pahole --format_path=btf -J --btf_encode_detached=/tmp/b \
	lib/ratelimit.o lib/radix-tree.o

combines gcc-generated BTF from lib/ratelimit.o and lib/radix-tree.o
into the /tmp/b detached BTF representation.

This is not an end in itself; it is clearly not optimal to
do this across several thousand kernel .o files, but it does
hopefully help pave the way to using BTF as input to pahole as is
proposed in [1].

[1] https://lore.kernel.org/dwarves/87ldqf1i19.fsf@esperi.org.uk/

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>