btf_encoder: Use addresses in debug info to identify .part, .cold functions
Because debug information does not include .part.0 or .cold.0 suffixes
we struggle to match the debug info for such functions with their
ELF function info which does contain such suffixes.

This is a problem because such debug info is irrelevant to checking
consistency of the "real" functions they are associated with, and yet
when we merge function info looking for inconsistencies they play
a(n unwanted) part.  So here we identify such cases by matching
the address in debug info (in this case DWARF) with the ELF function;
if we find such a match we skip saving function information for later
merging and thus sidestep this issue.

A concrete example of this is the function objpool_free().  It has
two entries in kallsyms/System.map:

ffffffff820409b0 t objpool_free.part.0
ffffffff82040a20 T objpool_free

We only care about the second one, but without identifying which
debug info is assocaited with objpool_free.part.0 we might falsely
flag objpool_free as inconsistent across representations.  But by
using the address in the debug info we can avoid this.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>