btf_encoder: Support encoding of inline location information
This patch requires updated libbpf with APIs to add location
information.  Iterate over inline expansions saving prototype
and associated location info for later addition.  Location
info can either be added to .BTF or a new .BTF.extra section
which is split BTF relative to .BTF; this helps when size of
location info makes adding it to .BTF prohibitive.  To support
this we need to dedup .BTF first, then access the mappings of
types to ensure the types of the parameters and return values
of the functions associated with the inline sites get post-dedup
updates.  Finally the .BTF.extra section itself is deduplicated
allowing for FUNC_PROTO, LOC_PARAM and LOC_PROTO deduplication.

Multiple BTF_KIND_LOCSECs are added if there are more then
65535 (max value vlen can support).

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>