dwarf_loader: share inlined value from abstract origin
When testing [1], it was observed that a function was showing
the wrong signature in BTF, namely

__be32 ip6_make_flowlabel(struct net *, struct sk_buff *, __be32, struct flowi6 *, bool);

The last two parameters are incorrectly reversed.  pfunct shows the right
thing for DWARF however.  It turns out the function is not always inlined;
a .constprop variant exists, and because of this _and_ the fact that
we do not share the inlined value between function representation and
its abstract origin, we incorrectly assume that function representation
(with abstract origin ref to the inlined function) is a "real" function
and use its incorrect signature.  So instead share the inlined value
to avoid this problem.

[1] https://lore.kernel.org/dwarves/20251003173620.2892942-1-yonghong.song@linux.dev/

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>